More Changes to Dtos, Forms Completed, might change the info on them (Names instead of ids, or maybe both)

+++ API/Core/Domain/Entities/Comment.cs
@@ -11,7 +11,7 @@ namespace Domain.Entities
     {
         public string Title { get; set; }
         public string Text { get; set; }
-        public int Rating { get; set; }
+        public byte Rating { get; set; }
 
         //// Relations
