+++ API/Infrastructure/Infrastructure/Dtos/Request/Comment/CommentUpdateVm.cs
@@ -6,11 +6,11 @@ using System.Threading.Tasks;
 
 namespace Infrastructure.Dtos.Request
 {
-   public class CommentUpdateVm
-   {
-      public int Id { get; set; }
-      public string? Title { get; set; }
-      public string? Text { get; set; }
-      public int? Rating { get; set; }
-   }
+    public class CommentUpdateVm
+    {
+        public int Id { get; set; }
+        public string? Title { get; set; }
+        public string? Text { get; set; }
+        public byte? Rating { get; set; }
+    }
 }
